Fix some renaming issues
